requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
flask-migrate
flask-session
flask-script
gevent
pandas
openpyxl
playwright
werkzeug
Flask-Mail
deep-translator
gunicorn
python-dotenv
XlsxWriter
flask-login
numpy
Babel
Flask-Babel
flask-jwt-extended